requirements.txt
Flask
python-dotenv
openai
google-generativeai
requests
PyMuPDF
beautifulsoup4
lxml
redis
rq
pandas
numpy
markdown
# gunicorn is for Linux deployments; on Windows use waitress or run via python
gunicorn

yfinance
flask_cors
chromadb
weasyprint

# Backend v2.1 Dependencies
fastapi>=0.110.0
pydantic>=2.0.0
sse-starlette>=3.0.0
uvicorn>=0.30.0
structlog>=24.0.0
flask_limiter
